requirements.txt
streamlit
google-generativeai
openai
pillow
matplotlib
PyPDF2
